+++ PyTM/__init__.py
@@ -1,3 +1,3 @@
 __author__ = "Wasi"
 __email__ = "wasi0013@gmail.com"
-__version__ = "0.0.8"
+__version__ = "0.0.9"